requirements.txt
transformers
torch
gradio
openai-whisper
bitsandbytes
pyyaml
fastapi
uvicorn

